still cleaning stuff up.
found a chunk of confidence code for tesseract results
   on google's tesseract page. need to do some tests.

added a filename namespace and a function to generate
   filenames based on the time of the robot. takes
   two arguments, a path and an extension.

+++ src/Navigation.h
@@ -5,8 +5,10 @@
 
 namespace navigation {
 
+	// this function translates the provided cell number
+	//    into a cardinal based on the last position
 	void add_to_path(char map[], const int cell) {
-		short last_post = map[0];
+		short last_post = (short)map[0];
 		short distance  = last_post - cell;
 
 		switch(distance) {
